fixes from comments - use translations and theme styling
