+++ common/config/rush/repo-state.json
@@ -1,5 +1,5 @@
 // DO NOT MODIFY THIS FILE MANUALLY BUT DO COMMIT IT. It is generated and used by Rush.
 {
-  "pnpmShrinkwrapHash": "8ee3968ee456957dd002890e619686a3597719ee",
+  "pnpmShrinkwrapHash": "f84ebd663b68ddc35b1ca0a226de3b73703576be",
   "preferredVersionsHash": "bf21a9e8fbc5a3846fb05b4fa0859e0917b2202f"
 }
