+++ packages/apps/transfer/src/components/Global/OptionsModal/OptionsModal.tsx
@@ -17,11 +17,17 @@ import {
   radioItemWrapperStyle,
   titleTagStyle,
 } from '@/components/Global/OptionsModal/styles.css';
-import { devOptions } from '@/constants/dev-options';
 import { DevOption } from '@/constants/kadena';
 import { useAppContext } from '@/context/app-context';
 import useTranslation from 'next-translate/useTranslation';
 import React, { FC, useState } from 'react';
+
+export interface IDevOption {
+  title: string;
+  text: string;
+  icon: React.FC;
+  tag?: string;
+}
 
 export const OptionsModal: FC = () => {
   const { t } = useTranslation('common');
@@ -29,6 +35,34 @@ export const OptionsModal: FC = () => {
   const [selected, setSelected] = useState(devOption);
 
   const { clearModal } = useModal();
+
+  const devOptions: {
+    [Key in DevOption]: IDevOption;
+  } = {
+    BASIC: {
+      title: t('Basic User Interface'),
+      text: t(
+        'This view can be used for basic operations and handling things within the user interface.',
+      ),
+      icon: SystemIcon.Application,
+    },
+    BACKEND: {
+      title: t('Backend Developers'),
+      tag: t('for PACT developers'),
+      text: t(
+        'This option is meant for developers who need more sophisticated options within the user interface.',
+      ),
+      icon: SystemIcon.ApplicationBrackets,
+    },
+    DAPP: {
+      title: t('dApp Developers'),
+      tag: t('for Javascript developers'),
+      text: t(
+        'This option is meant for developers who need more sophisticated options within the user interface.',
+      ),
+      icon: SystemIcon.ApplicationBrackets,
+    },
+  };
 
   const options = Object.entries(devOptions);
 
@@ -95,7 +129,7 @@ export const OptionsModal: FC = () => {
 
         <div className={modalButtonStyle}>
           <Button.Root
-            title="Save"
+            title={`${t('Save')}`}
             onClick={() => handleSave()}
             color="primary"
           >
